[NUI] CanvasView: Change parameter and return type of binder function
int -> uint(unsigned int)
dali-csharp-binder patch : https://review.tizen.org/gerrit/#/c/platform/core/uifw/dali-csharp-binder/+/262968/

+++ src/Tizen.NUI/src/public/BaseComponents/VectorGraphics/Gradient.cs
@@ -69,8 +69,8 @@ namespace Tizen.NUI.BaseComponents.VectorGraphics
             get
             {
                 List<ColorStop> retList = new List<ColorStop>();
-                int colorStopsCount = Interop.Gradient.GetColorStopsCount(BaseHandle.getCPtr(this));
-                for (int i = 0; i < colorStopsCount; i++)
+                uint colorStopsCount = Interop.Gradient.GetColorStopsCount(BaseHandle.getCPtr(this));
+                for (uint i = 0; i < colorStopsCount; i++)
                 {
                     retList.Add(new ColorStop(Interop.Gradient.GetColorStopsOffsetIndexOf(BaseHandle.getCPtr(this), i),
                                               Vector4.GetVector4FromPtr(Interop.Gradient.GetColorStopsColorIndexOf(BaseHandle.getCPtr(this), i))));
@@ -95,7 +95,7 @@ namespace Tizen.NUI.BaseComponents.VectorGraphics
                     rawColorStops[i * 5 + 4] = value[i].Color.A;
                 }
 
-                Interop.Gradient.SetColorStops(BaseHandle.getCPtr(this), rawColorStops, value.Count);
+                Interop.Gradient.SetColorStops(BaseHandle.getCPtr(this), rawColorStops, (uint)value.Count);
                 if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             }
         }
